+++ src/runtime/Tests/ArrayTests.m
@@ -26,8 +26,8 @@
 {
 	const long kValues[3] = { 1, 2, 3 };
 	PBArray *array = [[PBArray alloc] initWithValues:kValues count:3 valueType:PBArrayValueTypeInt32];
-	STAssertEquals([array count], (unsigned long)3, nil);
-	STAssertEquals(array.count, (unsigned long)3, nil);
+	STAssertEquals([array count], (NSUInteger)3, nil);
+	STAssertEquals(array.count, (NSUInteger)3, nil);
 	[array release];
 }
 
@@ -58,7 +58,7 @@
 - (void)testEmpty
 {
 	PBArray *array = [[PBArray alloc] init];
-	STAssertEquals(array.count, (unsigned long)0, nil);
+	STAssertEquals(array.count, (NSUInteger)0, nil);
 	STAssertEquals(array.valueType, PBArrayValueTypeObject, nil);
 	STAssertEquals(array.data, (const void *)nil, nil);
 	STAssertThrowsSpecificNamed([array objectAtIndex:0], NSException, NSRangeException, nil);
@@ -176,7 +176,7 @@
 	PBAppendableArray *array = [[PBAppendableArray alloc] initWithValueType:PBArrayValueTypeInt32];
 	[array addInt32:1];
 	[array addInt32:4];
-	STAssertEquals(array.count, (unsigned long)2, nil);
+	STAssertEquals(array.count, (NSUInteger)2, nil);
 	STAssertThrowsSpecificNamed([array addBool:NO], NSException, PBArrayTypeMismatchException, nil);
 	[array release];
 }
@@ -186,7 +186,7 @@
 	PBAppendableArray *array = [[PBAppendableArray alloc] initWithValueType:PBArrayValueTypeObject];
 	[array addObject:[NSString stringWithFormat:@"One"]];
 	[array addObject:[NSString stringWithFormat:@"Two"]];
-	STAssertEquals(array.count, (unsigned long)2, nil);
+	STAssertEquals(array.count, (NSUInteger)2, nil);
 	STAssertThrowsSpecificNamed([array addBool:NO], NSException, PBArrayTypeMismatchException, nil);
 	[array release];
 }
@@ -208,7 +208,7 @@
 	const long kValues[3] = { 1, 2, 3 };
 	PBAppendableArray *array = [[PBAppendableArray alloc] initWithValueType:PBArrayValueTypeInt32];
 	[array appendValues:kValues count:3];
-	STAssertEquals(array.count, (unsigned long)3, nil);
+	STAssertEquals(array.count, (NSUInteger)3, nil);
 	STAssertEquals([array int32AtIndex:1], (long)2, nil);
 	[array release];
 }
